:boom: Upgrade supported Vim version to 8.2.3452
The version exists in vim-win32-installer and closest smaller version that MacVim supports.

For Linux users
Vim installed with OS package managers are usually too old (e.g. 8.2.2434 in Ubuntu 21.04.)
That's why we need to force users to build it by themselves (and that what usual Vimmers do in Linux.)

For macOS users
Vim installed with Homebrew is almost HEAD but MacVim only support 8.2.3455 yet.
We need to make 8.2.3455 is the minimum supported version if we need to care MacVim.

For Windows users
Vim in https://github.com/vim/vim-win32-installer are quite close to the latest but some versions are
missed thus we need to find closest version if exact versions are missed. The closest version exists
for 8.2.3455 is 8.2.3452 (https://github.com/vim/vim-win32-installer/releases/tag/v8.2.3452)