:boom: Upgrade supported Neovim version to 0.9.4
The most recent version of Neovim is 0.9.5, so we provide support
for version 0.9.4 and newer. Please note that while we mention
supporting the two most recent minor versions of Neovim, the
software has not reached version 1.0. Therefore, we consider the
patch version to be equivalent to the minor version for the
purpose of support.