:boom: Remove `@denops` module and use `denops_core` instead
Previously, we authored the public interface of denops within the
`@denops` module and released it to deno.land as `denops_core`.
However, maintaining both the type definitions and implementations in
the same repository proved challenging.
Consequently, we opted to split them into two separate repositories, with
the type definitions now residing in the `deno-denops-core` repository.

https://github.com/vim-denops/deno-denops-core