:boom: Remove `Worker` and load plugins in main thread
Denops has used a worker as a plugin isolation mechanism since
its development, but found that as the number of plugins
increased, non-negligible performance problems occurred because
the startup of a worker takes several milliseconds.

Also, while many plugins have been created since Denops was
introduced, none have had such a negative impact on other plugins
that they need to be separated and managed.

From the above, we decided that the separation mechanism using a
worker has more disadvantages, and decided to change to loading
each plugin into the main thread.

This change may affect other plugins if any plugins are performing
heavy processing, but we believe that in such a case, each plugin
can minimize the impact by setting up a separate worker and letting
the processing escape.